refactor: renaming the files from todo to tasks
