--- src/tasks/todo.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TodoController } from './todo.controller';
-import { TodoService } from './todo.service';
-
-@Module({
-  controllers: [TodoController],
-  providers: [TodoService]
-})
-export class TodoModule {}
